+++ R/simulation-simInit.R
@@ -1461,8 +1461,9 @@ simInitAndSpades <- function(times, params, modules, objects, paths, inputs, out
 
           # nextEvent <- NULL
           runFnCallAsExpr <- TRUE
+          debugForCache <- debugToVerbose(debug)
           if (allowSequentialCaching) {
-            sim <- allowSequentialCaching1(sim, cacheIt, moduleCall = ".inputObjects", verbose = debug)
+            sim <- allowSequentialCaching1(sim, cacheIt, moduleCall = ".inputObjects", verbose = debugForCache)
             runFnCallAsExpr <- is.null(attr(sim, "runFnCallAsExpr"))
           }
           if (runFnCallAsExpr) {
@@ -1470,7 +1471,6 @@ simInitAndSpades <- function(times, params, modules, objects, paths, inputs, out
             pkgs <- c(pkgs, "stats")
             # if (getOption("spades.useBox", FALSE) && FALSE)
             #   do.call(box::use, lapply(pkgs, as.name))
-            debugForCache <- debugToVerbose(debug)
             if (any(mBase %in% getOption("spades.debugModule"))) {
               browser()
             }