+++ NEWS.md
@@ -10,6 +10,9 @@
 * several corrections to messaging when there is nested simLists, i.e., when simInit is called within a module; 
 * new function `doCallSafe` that can be used for `doCallSafe(simInitAndSpades, out)`, and it does not suffer from the slow downs of `do.call`; 
 * many changes to accommodate updates to `reproducible` package, specifically the `Cache` function. These include
+* fix warning capture in `spades()` debug logging;
+* fix use of `debug` as `verbose` when caching (#322);
+* improve startup message formatting;
 * documentation improvements;
 
 # SpaDES.core 2.1.8